requirements.txt
discord.py[voice]>=1.4.1
googletrans==4.0.0-rc1